add util for better organization
